sipXconfig/neoconf/test/org/sipfoundry/sipxconfig/setting/ModelBuilderCacheInterceptorTest.java
/*
 *
 *
 * Copyright (C) 2007 Pingtel Corp., certain elements licensed under a Contributor Agreement.
 * Contributors retain copyright to elements licensed under a Contributor Agreement.
 * Licensed to the User under the LGPL license.
 *
 * $
 */
package org.sipfoundry.sipxconfig.setting;

import java.io.File;
import java.io.Serializable;

import junit.framework.TestCase;
import net.sf.ehcache.Cache;
import net.sf.ehcache.Element;

import org.easymock.EasyMock;
import org.easymock.IArgumentMatcher;
import org.easymock.IMocksControl;
import org.springframework.aop.framework.ProxyFactory;

public class ModelBuilderCacheInterceptorTest extends TestCase {

    // FIXME: REQUIRED FOR PORTING TO SPRING 2.0
    public void testNop() {
    }

    // FIXME: FAILED WHILE PORTING TO SPRING 2.0
    public void DISABLED_testInterceptor() throws Exception {
        Setting abc = new SettingSet("abc");
        Setting cde = new SettingSet("cde");

        File file1 = new File("abc");
        File file2 = new File("cde");
        File file3 = new File("abc");

        Element el1 = new Element(file1.getPath(), (Serializable) abc);

        IMocksControl cacheControl = org.easymock.classextension.EasyMock.createControl();
        Cache cache = cacheControl.createMock(Cache.class);
        cache.get(file1.getPath());
        cacheControl.andReturn(null);
        cache.put(ElementEquals.matches("abc"));
        cache.get(file2.getPath());
        cacheControl.andReturn(null);
        cache.put(ElementEquals.matches("cde"));
        cache.get(file3.getPath());
        cacheControl.andReturn(el1);

        cacheControl.replay();

        IMocksControl modelBuilderControl = EasyMock.createControl();
        ModelBuilder modelBuilder = modelBuilderControl.createMock(ModelBuilder.class);
        modelBuilder.buildModel(file1);
        modelBuilderControl.andReturn(abc);
        modelBuilder.buildModel(file2);
        modelBuilderControl.andReturn(cde);
        modelBuilderControl.replay();

        ProxyFactory proxyFactory = new ProxyFactory(modelBuilder);
        ModelBuilderCacheInterceptor interceptor = new ModelBuilderCacheInterceptor();
        interceptor.setCache(cache);
        proxyFactory.addAdvice(interceptor);

        ModelBuilder proxy = (ModelBuilder) proxyFactory.getProxy();
        assertSame(abc, proxy.buildModel(file1));
        assertSame(cde, proxy.buildModel(file2));
        assertSame(abc, proxy.buildModel(file3));

        modelBuilderControl.verify();
        cacheControl.verify();
    }

    /** Match all elements with the same key */
    private static class ElementEquals implements IArgumentMatcher {
        public static Element matches(String key) {
            EasyMock.reportMatcher(new ElementEquals(key));
            return null;
        }

        private String m_key;

        public ElementEquals(String key) {
            m_key = key;
        }

        public void appendTo(StringBuffer buffer) {
            buffer.append(m_key);
        }

        public boolean matches(Object argument) {
            Element el = (Element) argument;
            return el.getKey().equals(m_key);
        }
    }
}
